requirements.txt
aws-cdk-lib==2.189.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-python-alpha==2.175.1a0
cdklabs.generative-ai-cdk-constructs==0.1.298
boto3==1.36.16
